Move function implementation
